Restructured controllers to match new API model
